refactor code
move files to standardize them
