+++ app/javascript/gobierto_visualizations/webapp/containers/contracts/AssigneesShow.vue
@@ -27,7 +27,7 @@
 <script>
 import { VueFiltersMixin } from "lib/vue/filters"
 import { Table } from "lib/vue/components";
-import { EventBus } from "../../mixins/event_bus";
+import { EventBus } from "../../lib/mixins/event_bus";
 import { assigneesShowColumns } from "../../lib/config/contracts.js";
 
 export default {
