+++ app/javascript/gobierto_visualizations/webapp/containers/contracts/Home.vue
@@ -32,8 +32,8 @@ import ContractsIndex from "./ContractsIndex.vue";
 import ContractsShow from "./ContractsShow.vue";
 import AssigneesShow from "./AssigneesShow.vue";
 
-import { EventBus } from "../../mixins/event_bus";
-import { store } from "../../mixins/store";
+import { EventBus } from "../../lib/mixins/event_bus";
+import { store } from "../../lib/mixins/store";
 
 export default {
   name: 'Home',
