+++ app/javascript/gobierto_visualizations/webapp/containers/costs/Home.vue
@@ -53,7 +53,7 @@
 <script>
 import Distribution from './Distribution.vue'
 import Table from './table/Table.vue'
-import { EventBus } from "../../mixins/event_bus";
+import { EventBus } from "../../lib/mixins/event_bus";
 
 export default {
   name: 'Home',
