+++ app/javascript/gobierto_visualizations/webapp/containers/subsidies/Aside.vue
@@ -48,7 +48,7 @@
 import { BlockHeader, Checkbox, Dropdown } from "lib/vue/components";
 import DownloadButton from "../../components/DownloadButton.vue";
 import SearchFilter from "../../components/SearchFilter.vue";
-import { EventBus } from "../../mixins/event_bus";
+import { EventBus } from "../../lib/mixins/event_bus";
 import { subsidiesFiltersConfig } from "../../lib/config/subsidies.js";
 
 export default {
