+++ app/javascript/gobierto_visualizations/webapp/containers/subsidies/CategoriesTreeMapNested.vue
@@ -26,7 +26,7 @@
 <script>
 import { nest } from "d3-collection";
 import { select, mouse } from 'd3-selection';
-import TreeMapNested from "../../visualizations/treeMapNested.vue";
+import TreeMapNested from "../../components/TreeMapNested.vue";
 import { money } from "lib/vue/filters";
 
 const d3 = { nest, select, mouse }
