+++ app/javascript/gobierto_visualizations/webapp/lib/mixins/event_bus.js
