+++ app/javascript/gobierto_visualizations/webapp/lib/mixins/store.js
