+++ app/javascript/lib/vue/components/index.js
@@ -8,7 +8,6 @@ import ReadMore from "./modules/ReadMore.vue"
 import Dropdown from "./modules/Dropdown.vue"
 import SkeletonSpinner from "./modules/SkeletonSpinner.vue"
 import Pagination from "./modules/Pagination.vue"
-import BeesWarmChart from "./modules/BeesWarmChart.vue"
 import MultipleLineChart from "./modules/MultipleLineChart.vue"
 import TextEditable from "./modules/TextEditable.vue"
 import Autocomplete from "./modules/Autocomplete.vue"
@@ -28,7 +27,6 @@ export {
   Dropdown,
   SkeletonSpinner,
   Pagination,
-  BeesWarmChart,
   MultipleLineChart,
   TextEditable,
   Autocomplete,
